requirements.txt
pydantic
pyYAML
pyexcel
pyexcel-xls
pygame
watchdog
pysftp
xlrd
jinja2
imgkit
pytest
pywin32
